+++ angular-workspace/projects/ni/nimble-angular/src/directives/anchor-tree-item/nimble-anchor-tree-item.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { NimbleAnchorTreeItemDirective } from './nimble-anchor-tree-item.directive';
+import { NimbleAnchorTreeItemRouterLinkDirective } from './nimble-anchor-tree-item-router-link.directive';
+import { NimbleAnchorTreeItemRouterLinkWithHrefDirective } from './nimble-anchor-tree-item-router-link-with-href.directive';
+
+import '@ni/nimble-components/dist/esm/anchor-tree-item';
+
+@NgModule({
+    declarations: [NimbleAnchorTreeItemDirective, NimbleAnchorTreeItemRouterLinkDirective, NimbleAnchorTreeItemRouterLinkWithHrefDirective],
+    imports: [CommonModule],
+    exports: [NimbleAnchorTreeItemDirective, NimbleAnchorTreeItemRouterLinkDirective, NimbleAnchorTreeItemRouterLinkWithHrefDirective]
+})
+export class NimbleAnchorTreeItemModule { }