+++ packages/nimble-components/src/anchor-tab/index.ts
@@ -30,6 +30,8 @@ export class AnchorTab extends AnchorBase {
     public disabled = false;
 }
 
+// FoundationAnchor already applies the StartEnd mixin, so we don't need to do it here.
+
 const nimbleAnchorTab = AnchorTab.compose({
     baseName: 'anchor-tab',
     template,