+++ packages/nimble-components/src/table/index.ts
@@ -38,6 +38,7 @@ import { Virtualizer } from './models/virtualizer';
 import { getTanStackSortingFunction } from './models/sort-operations';
 import { UpdateTracker } from './models/update-tracker';
 import { TableLayoutHelper } from './models/table-layout-helper';
+import type { TableRow } from './components/row';
 
 declare global {
     interface HTMLElementTagNameMap {
@@ -85,6 +86,12 @@ export class Table<
      * @internal
      */
     @observable
+    public readonly rowElements: TableRow[] = [];
+
+    /**
+     * @internal
+     */
+    @observable
     public actionMenuSlots: string[] = [];
 
     /**
@@ -119,6 +126,11 @@ export class Table<
      * @internal
      */
     public readonly viewport!: HTMLElement;
+
+    /**
+     * @internal
+     */
+    public readonly rowContainer!: HTMLElement;
 
     /**
      * @internal
@@ -285,6 +297,9 @@ export class Table<
         event.stopImmediatePropagation();
 
         this.$emit('action-menu-toggle', event.detail);
+        if (!event.detail.newState) {
+            this.openActionMenuRecordId = undefined;
+        }
     }
 
     /**